requirement.txt
django==1.10
djangorestframework
django-cors-headers
mysql-python